+++ examples/resources/gigahost_server_reverse/resource.tf
@@ -1,4 +1,4 @@
-resource "gigahost_server_ip_reverse" "example" {
+resource "gigahost_server_reverse" "example" {
   srv_id     = gigahost_server.example.srv_id
   ip_id      = 2002
   ip_v4v6    = "ipv4"